requirements.txt
fastapi
uvicorn
supabase
langchain-groq
python-dotenv
pendulum